refactor(translation): drop -core suffix, name storage key by language
Rename board-translation-core to board-translation now that the use-
prefix already distinguishes the hook. Rename the updateBoardStrings /
persistTranslations key param from locale to language: callers pass a
base language code (e.g. "es"), not a full locale, so the name now
matches the value and the language-keyed storage convention is explicit.
