+++ src/features/board/storage/db.ts
@@ -253,7 +253,7 @@ export async function getBoard(
 export async function updateBoardStrings(
   setId: string,
   boardId: string,
-  locale: string,
+  language: string,
   translations: Record<string, string>,
 ): Promise<void> {
   validateId(setId, "setId");
@@ -268,7 +268,7 @@ export async function updateBoardStrings(
 
   const updatedObf = {
     ...record.obf,
-    strings: { ...record.obf.strings, [locale]: translations },
+    strings: { ...record.obf.strings, [language]: translations },
   };
 
   await tx.store.put({ ...record, obf: updatedObf });
